refactor(recordings): decompose recording detail page into data/live/transcript/notes/documents/action modules (FE-011)
